chore(server): improve naming of loadout components and move into a folder
